added mkdir

+++ debInstall32Bit.bash
@@ -10,6 +10,7 @@ then
     die "$0: Received $# args... version and whether or not this is a release required"
 fi
 
-cp lib/linux/x86/libunix-java.so install/linux/lib  || die "Could not copy libunix?"
+test -d install/linux/lib || mkdir -p install/linux/lib || die "Could not create install/linux/lib"
+cp lib/linux/x86/libunix-java.so install/linux/lib/  || die "Could not copy libunix?"
 ./debInstall.bash $* 32 690
 rm install/linux/lib/*
